+++ ChangeLog.md
@@ -1,4 +1,5 @@
 ### 5.1.2 (TBD)
+* Fix issue where extracting a string from a DICOM dataset could return null if the tag was present but empty
 * Add the possibility to configure TCP buffer sizes (#1564)
 
 #### 5.1.1 (2023-05-29)